Moved commonly used padding method from SecureMessaging to CryptoUtil
